Clean up outbound init
