+++ telepresence/outbound/local.py
@@ -124,3 +124,50 @@ def launch_vpn(
         "Terminate local process", terminate_local_process, runner, process
     )
     return process
+
+
+def setup_inject(runner: Runner, args):
+    runner.require(["torsocks"], "Please install torsocks (v2.1 or later)")
+    if runner.chatty:
+        runner.show(
+            "Starting proxy with method 'inject-tcp', which has the following "
+            "limitations: Go programs, static binaries, suid programs, and "
+            "custom DNS implementations are not supported. For a full list of "
+            "method limitations see "
+            "https://telepresence.io/reference/methods.html"
+        )
+    command = ["torsocks"] + (args.run or ["bash", "--norc"])
+
+    def launch(runner_, _remote_info, env, socks_port, _ssh, _mount_dir):
+        return launch_inject(runner_, command, socks_port, env)
+
+    return launch
+
+
+def setup_vpn(runner: Runner, args):
+    runner.require(["sshuttle-telepresence"],
+                   "Part of the Telepresence package. Try reinstalling.")
+    if runner.platform == "linux":
+        # Need conntrack for sshuttle on Linux:
+        runner.require(["conntrack", "iptables"],
+                       "Required for the vpn-tcp method")
+    if runner.platform == "darwin":
+        runner.require(["pfctl"], "Required for the vpn-tcp method")
+    runner.require_sudo()
+    if runner.chatty:
+        runner.show(
+            "Starting proxy with method 'vpn-tcp', which has the following "
+            "limitations: All processes are affected, only one telepresence "
+            "can run per machine, and you can't use other VPNs. You may need "
+            "to add cloud hosts and headless services with --also-proxy. For "
+            "a full list of method limitations see "
+            "https://telepresence.io/reference/methods.html"
+        )
+    command = args.run or ["bash", "--norc"]
+
+    def launch(runner_, remote_info, env, _socks_port, ssh, _mount_dir):
+        return launch_vpn(
+            runner_, remote_info, command, args.also_proxy, env, ssh
+        )
+
+    return launch
